subscriber2/requirements.txt
azure-functions
azure-servicebus
azure-cosmos